requirements.txt
Pillow
tk
opencv-python
